実装終わり

+++ Core/Inc/ICM20948_USER.h
@@ -9,7 +9,7 @@
 #define ICM20948_ICM20948_USER_H_
 
 #include "ICM20948/ICM20948_HAL.h"
-#include "Quaternion/Vector3D/Vector3D.h"
+#include "Vector3D/Vector3D.h"
 #include "stdexcept"
 #include <array>
 
@@ -28,14 +28,13 @@ public:
 		return __isCalibrated;
 	}
 
-	void getIMU(Vector3D<float> &accel, Vector3D<float> &gyro);
+	void getIMU(Vector3D<int16_t> &accel, Vector3D<int16_t> &gyro);
 
 	void update(){};
 
 private:
 	ICM20948_HAL *icm20948;
 	bool __isCalibrated;
-//	Vector3D<double> accelAverage={-0.059056974459725, -0.0230255402750491, -0.0633869208693285};
 	Vector3D<double> accelAverage={-0.04,-0.05,-0.0633};
 	Vector3D<double> gyroAverage={};
 	uint16_t averageCounter;
